Restructured the GUI folder and added a GPIO pending IRQ column (visualization)
