+++ src/gui/windows/peripherals/ram_window.cpp
